(hand)Written rendering pipeline
This update explores the idea of being able to follow a CG graphics rendering pipeline by hand with only a good calculator to be able to at least come up with an outline drawing that can be rotated and viewed from different perspective angles and maintain proportions and allow for further hand painted work on top of this outline. By pressing the Written Render button and reading the code, the user follows a series of steps that can be solved with only a calculator.