gradient method 0 more procedural
gradient method 0 now has procedural controls for setting step, row range, size of color swatches, spacing, and height.

Additionally, the color wheel preset entitled "color wheel preset 18 0" is recommended because it splits the color spectrum in 18 distinct hues. This preset can also be printed out on glossy paper and then trimmed between the hue strings and attached to something straw shaped in the middle with tape and then be used to measure colors of objects behind this pinwheel like device and identify them. There is an 'X' printed on every 5th swatch and can be used to match up to a second printing of the color wheel that is laminated.